refactored scoop badges - added scoop base
