+++ services/scoop/scoop-license.service.js
@@ -1,19 +1,10 @@
-import { URL } from 'url'
 import Joi from 'joi'
-import { NotFound, pathParam, queryParam } from '../index.js'
-import { ConditionalGithubAuthV3Service } from '../github/github-auth-service.js'
-import { fetchJsonFromRepo } from '../github/github-common-fetch.js'
+import { pathParam, queryParam } from '../index.js'
 import { renderLicenseBadge } from '../licenses.js'
 import toArray from '../../core/base-service/to-array.js'
+import ScoopBase from './scoop-base.js'
 
-const gitHubRepoRegExp =
-  /https:\/\/github.com\/(?<user>.*?)\/(?<repo>.*?)(\/|$)/
-
-const bucketsSchema = Joi.object()
-  .pattern(/.+/, Joi.string().pattern(gitHubRepoRegExp).required())
-  .required()
-
-const scoopSchema = Joi.object({
+const scoopLicenseSchema = Joi.object({
   license: Joi.alternatives()
     .try(
       Joi.string().required(),
@@ -28,7 +19,7 @@ const queryParamSchema = Joi.object({
   bucket: Joi.string(),
 })
 
-export default class ScoopLicense extends ConditionalGithubAuthV3Service {
+export default class ScoopLicense extends ScoopBase {
   // The buckets file (https://github.com/lukesampson/scoop/blob/master/buckets.json) changes very rarely.
   // Cache it for the lifetime of the current Node.js process.
   buckets = null
@@ -67,68 +58,15 @@ export default class ScoopLicense extends ConditionalGithubAuthV3Service {
   }
 
   async handle({ app }, queryParams) {
-    if (!this.buckets) {
-      this.buckets = await fetchJsonFromRepo(this, {
-        schema: bucketsSchema,
-        user: 'ScoopInstaller',
-        repo: 'Scoop',
-        branch: 'master',
-        filename: 'buckets.json',
-      })
-    }
+    const { license } = await this.fetch(
+      { app, schema: scoopLicenseSchema },
+      queryParams,
+    )
 
-    const bucket = queryParams.bucket || 'main'
-    let bucketUrl = this.buckets[bucket]
+    const licenses = toArray(license).map(license =>
+      typeof license === 'string' ? license : license.identifier,
+    )
 
-    if (!bucketUrl) {
-      // Parsing URL here will throw an error if the url is invalid
-      try {
-        const url = new URL(decodeURIComponent(bucket))
-
-        // Throw errors to go to jump to catch statement
-        // The error messages here are purely for code readability, and will never reach the user.
-        if (url.hostname !== 'github.com') {
-          throw new Error('Not a GitHub URL')
-        }
-        const path = url.pathname.split('/').filter(value => value !== '')
-
-        if (path.length !== 2) {
-          throw new Error('Not a valid GitHub Repo')
-        }
-
-        const [user, repo] = path
-
-        // Reconstructing the url here ensures that the url will match the regex
-        bucketUrl = `https://github.com/${user}/${repo}`
-      } catch (e) {
-        throw new NotFound({ prettyMessage: `bucket "${bucket}" not found` })
-      }
-    }
-
-    const {
-      groups: { user, repo },
-    } = gitHubRepoRegExp.exec(bucketUrl)
-
-    try {
-      const { license } = await fetchJsonFromRepo(this, {
-        schema: scoopSchema,
-        user,
-        repo,
-        branch: 'master',
-        filename: `bucket/${app}.json`,
-      })
-
-      const licenses = toArray(license).map(license =>
-        typeof license === 'string' ? license : license.identifier,
-      )
-      return this.constructor.render({ licenses })
-    } catch (error) {
-      if (error instanceof NotFound) {
-        throw new NotFound({
-          prettyMessage: `${app} not found in bucket "${bucket}"`,
-        })
-      }
-      throw error
-    }
+    return this.constructor.render({ licenses })
   }
 }
